pyproject.toml
[project]
name = "team-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (==0.115.9)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "chromadb (>=1.0.10,<2.0.0)",
    "openai (>=1.82.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-openai (>=0.3.18,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
rag_agent = "uvicorn rag_agent.app:app --reload"